Allow store and recall of master fader
I'm not sure I'm totally happy with this, but here's how it works.

level is stored in *Piece*, not configuration. As BMMasterFader is part of the chain, we need to stop it being set there. (In fact this only didn't happen because of the bug fixed in efab0bc. For generic use we don't want to kill this.)

Solution is to add a 'acceptsMappings' var, which when false means mappings returns an empty dict, and mappings_ is ignored, just posting a warning if a numeric value is stored. (You can avoid this warning in BMSystem byt restoring the config.)

A separate loadPiece method sets the level from a pieceEvent, a la sound file player.

Master fader level is stored along with all controllers.

Signed-off-by: Scott Wilson <i@scottwilson.ca>