Fix for problem when importing concert
chainFunc didn't build the SoundFilePlayer again, so this meant that when asBMInOutArray was called on the old one, we'd get an error since its bus had been freed.

I can't recall offhand why we need to rebuild the chain every time we import a concert, but that's another question (maybe for Sergio).